feat: remove user-home offer, demand, swap page, add offer/[businessFunction] page
