--- app/user/[username]/(home)/deal/demand/layout.tsx
@@ -1,8 +1,0 @@
-import TextFeedLayout from '@/components/deals/text-feed.layout';
-import React from 'react';
-
-function MyDemandsLayout({ children }: { children: React.ReactNode }) {
-  return <TextFeedLayout>{children}</TextFeedLayout>;
-}
-
-export default MyDemandsLayout;
