--- app/user/[username]/(home)/deal/swap/layout.tsx
@@ -1,8 +1,0 @@
-import ThumbnailFeedLayout from '@/components/deals/thumbnail-feed.layout';
-import React from 'react';
-
-function MySwapsLayout({ children }: { children: React.ReactNode }) {
-  return <ThumbnailFeedLayout>{children}</ThumbnailFeedLayout>;
-}
-
-export default MySwapsLayout;
