+++ app/user/[username]/(home)/offer/page.tsx
@@ -5,5 +5,5 @@ import { redirect, usePathname } from 'next/navigation';
 export default function Page() {
   const pathname = usePathname();
 
-  return redirect(`${pathname}/offer`);
+  return redirect(`${pathname}/sell`);
 }
